Move DSA* interfaces to interfaces.dsa module
